+++ CHANGELOG.md
@@ -9,6 +9,12 @@ his file's format is define in
 and the release workflow reads it to set github's release notes.
 
 
+## Unreleased
+
+### Added
+
+- WHIP support
+
 ## [1.2.2] 2023-9-11
 
 ### Fixed
@@ -19,7 +25,11 @@ and the release workflow reads it to set github's release notes.
 
 ### Fixed 
 
+<<<<<<< HEAD
 - first marker restore 
+=======
+- first market restore 
+>>>>>>> f8c5dc2 (Releasing v1.2.1)
 
 ## [1.2.0] 2023-9-6